requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-rest-swagger
python-dotenv
psycopg2
asgiref
black
black
flake8
isort
pre-commit
factory_boy
django-cors-headers
drf_yasg
sentry-sdk
Pillow
python-logstash
django-elasticsearch-dsl
django-elasticsearch-dsl-drf
pbl6packageg2==0.0.1